Added some more test around basic foreach execution.
